requirements.txt
bleach==3.1.0
certifi==2019.9.11
chardet==3.0.4
Django==2.2.6
django-nose==1.4.6
djangorestframework==3.10.3
docutils==0.15.2
idna==2.8
nose==1.3.7
pkginfo==1.5.0.1
psycopg2==2.8.4
Pygments==2.4.2
PyJWT==1.7.1
python-decouple==3.1
pytz==2019.3
readme-renderer==24.0
requests==2.22.0
requests-toolbelt==0.9.1
six==1.12.0
sqlparse==0.3.0
tqdm==4.36.1
urllib3==1.25.6
webencodings==0.5.1
